requirements.txt
langchain
langchain-groq
langchain-openai
langchain-google-genai
langchain-anthropic
python-dotenv
pydantic
fastapi
uvicorn[standard]
groq
